# src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"HealthResponse": {
				"properties": {
					"status": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"timestamp": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DiagnosisResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"id",
					"code",
					"name",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ClinicResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateClinicRequest": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateClinicRequest": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ClinicianType": {
				"enum": [
					"doctor",
					"nurse",
					"therapist",
					"psychiatrist"
				],
				"type": "string"
			},
			"ClinicSummary": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ClinicianResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/ClinicianType"
					},
					"clinic": {
						"$ref": "#/components/schemas/ClinicSummary"
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"type",
					"clinic"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateClinicianRequest": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/ClinicianType"
					},
					"clinicId": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"type",
					"clinicId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateClinicianRequest": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/ClinicianType"
					},
					"clinicId": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"LoginResponse": {
				"properties": {
					"token": {
						"type": "string"
					},
					"expiresIn": {
						"type": "string"
					}
				},
				"required": [
					"token",
					"expiresIn"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginRequest": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "bluesky-aba-core",
		"version": "1.0.0",
		"description": "Clinical backend API for ABA therapy clinics built with Node.js, Express, TypeScript, PostgreSQL, TypeORM, Zod, and openFDA integration.",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/health": {
			"get": {
				"operationId": "GetHealth",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HealthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Health"
				],
				"security": [],
				"parameters": []
			}
		},
		"/diagnoses": {
			"get": {
				"operationId": "GetDiagnoses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DiagnosisResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Diagnoses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "code",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/diagnoses/{id}": {
			"get": {
				"operationId": "GetDiagnosis",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DiagnosisResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Diagnoses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/diagnoses/code/{code}": {
			"get": {
				"operationId": "GetDiagnosisByCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DiagnosisResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Diagnoses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/clinics": {
			"get": {
				"operationId": "GetClinics",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ClinicResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Clinics"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateClinic",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ClinicResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Clinics"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateClinicRequest"
							}
						}
					}
				}
			}
		},
		"/clinics/{id}": {
			"get": {
				"operationId": "GetClinic",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ClinicResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Clinics"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateClinic",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ClinicResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Clinics"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateClinicRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteClinic",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Clinics"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/clinicians": {
			"get": {
				"operationId": "GetClinicians",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ClinicianResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Clinicians"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "clinicId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ClinicianType"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateClinician",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ClinicianResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Clinicians"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateClinicianRequest"
							}
						}
					}
				}
			}
		},
		"/clinicians/{id}": {
			"get": {
				"operationId": "GetClinician",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ClinicianResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Clinicians"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateClinician",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ClinicianResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Clinicians"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateClinicianRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteClinician",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Clinicians"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Authorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}